# docs/api-docs.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Tutorial Management API",
		"description": "This API exposes endpoints to manage tutorials.",
		"contact": {
			"name": "Max Mustermann",
			"url": "https://www.MaxMustermann.com",
			"email": "max.mustermann@gmail.com"
		},
		"license": {
			"name": "The Unlicense",
			"url": "https://choosealicense.com/licenses/unlicense/"
		},
		"version": "1.0"
	},
	"servers": [
		{
			"url": "http://localhost:8081",
			"description": "Server URL in Development environment"
		},
		{
			"url": "https://to_be_defined.com",
			"description": "Server URL in Test environment"
		}
	],
	"tags": [
		{
			"name": "Tutorial",
			"description": "Tutorial management APIs"
		}
	],
	"paths": {
		"/api/tutorials/{id}": {
			"get": {
				"tags": [
					"Tutorial",
					"tutorials",
					"get"
				],
				"summary": "Find a Tutorial by Id",
				"description": "Returns a single tutorial by ID.",
				"operationId": "getTutorialById",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/TutorialDto"
								}
							}
						}
					},
					"404": {
						"description": "Tutorial not found",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/TutorialDto"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/TutorialDto"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": [
					"Tutorial",
					"tutorials",
					"put"
				],
				"summary": "Update a Tutorial by Id",
				"description": "Updates a tutorial based on its ID.",
				"operationId": "updateTutorial",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TutorialDto"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Tutorial successfully updated",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/TutorialDto"
								}
							}
						}
					},
					"404": {
						"description": "Tutorial not found",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/TutorialDto"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/TutorialDto"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": [
					"Tutorial",
					"tutorials",
					"delete"
				],
				"summary": "Delete a Tutorial by Id",
				"description": "Deletes the tutorial with the specified ID.",
				"operationId": "deleteTutorial",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Tutorial successfully deleted",
						"content": {
							"*/*": {
								"schema": {
									"type": "string",
									"enum": [
										"100 CONTINUE",
										"101 SWITCHING_PROTOCOLS",
										"102 PROCESSING",
										"103 EARLY_HINTS",
										"103 CHECKPOINT",
										"200 OK",
										"201 CREATED",
										"202 ACCEPTED",
										"203 NON_AUTHORITATIVE_INFORMATION",
										"204 NO_CONTENT",
										"205 RESET_CONTENT",
										"206 PARTIAL_CONTENT",
										"207 MULTI_STATUS",
										"208 ALREADY_REPORTED",
										"226 IM_USED",
										"300 MULTIPLE_CHOICES",
										"301 MOVED_PERMANENTLY",
										"302 FOUND",
										"302 MOVED_TEMPORARILY",
										"303 SEE_OTHER",
										"304 NOT_MODIFIED",
										"305 USE_PROXY",
										"307 TEMPORARY_REDIRECT",
										"308 PERMANENT_REDIRECT",
										"400 BAD_REQUEST",
										"401 UNAUTHORIZED",
										"402 PAYMENT_REQUIRED",
										"403 FORBIDDEN",
										"404 NOT_FOUND",
										"405 METHOD_NOT_ALLOWED",
										"406 NOT_ACCEPTABLE",
										"407 PROXY_AUTHENTICATION_REQUIRED",
										"408 REQUEST_TIMEOUT",
										"409 CONFLICT",
										"410 GONE",
										"411 LENGTH_REQUIRED",
										"412 PRECONDITION_FAILED",
										"413 PAYLOAD_TOO_LARGE",
										"413 REQUEST_ENTITY_TOO_LARGE",
										"414 URI_TOO_LONG",
										"414 REQUEST_URI_TOO_LONG",
										"415 UNSUPPORTED_MEDIA_TYPE",
										"416 REQUESTED_RANGE_NOT_SATISFIABLE",
										"417 EXPECTATION_FAILED",
										"418 I_AM_A_TEAPOT",
										"419 INSUFFICIENT_SPACE_ON_RESOURCE",
										"420 METHOD_FAILURE",
										"421 DESTINATION_LOCKED",
										"422 UNPROCESSABLE_ENTITY",
										"423 LOCKED",
										"424 FAILED_DEPENDENCY",
										"425 TOO_EARLY",
										"426 UPGRADE_REQUIRED",
										"428 PRECONDITION_REQUIRED",
										"429 TOO_MANY_REQUESTS",
										"431 REQUEST_HEADER_FIELDS_TOO_LARGE",
										"451 UNAVAILABLE_FOR_LEGAL_REASONS",
										"500 INTERNAL_SERVER_ERROR",
										"501 NOT_IMPLEMENTED",
										"502 BAD_GATEWAY",
										"503 SERVICE_UNAVAILABLE",
										"504 GATEWAY_TIMEOUT",
										"505 HTTP_VERSION_NOT_SUPPORTED",
										"506 VARIANT_ALSO_NEGOTIATES",
										"507 INSUFFICIENT_STORAGE",
										"508 LOOP_DETECTED",
										"509 BANDWIDTH_LIMIT_EXCEEDED",
										"510 NOT_EXTENDED",
										"511 NETWORK_AUTHENTICATION_REQUIRED"
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"*/*": {
								"schema": {
									"type": "string",
									"enum": [
										"100 CONTINUE",
										"101 SWITCHING_PROTOCOLS",
										"102 PROCESSING",
										"103 EARLY_HINTS",
										"103 CHECKPOINT",
										"200 OK",
										"201 CREATED",
										"202 ACCEPTED",
										"203 NON_AUTHORITATIVE_INFORMATION",
										"204 NO_CONTENT",
										"205 RESET_CONTENT",
										"206 PARTIAL_CONTENT",
										"207 MULTI_STATUS",
										"208 ALREADY_REPORTED",
										"226 IM_USED",
										"300 MULTIPLE_CHOICES",
										"301 MOVED_PERMANENTLY",
										"302 FOUND",
										"302 MOVED_TEMPORARILY",
										"303 SEE_OTHER",
										"304 NOT_MODIFIED",
										"305 USE_PROXY",
										"307 TEMPORARY_REDIRECT",
										"308 PERMANENT_REDIRECT",
										"400 BAD_REQUEST",
										"401 UNAUTHORIZED",
										"402 PAYMENT_REQUIRED",
										"403 FORBIDDEN",
										"404 NOT_FOUND",
										"405 METHOD_NOT_ALLOWED",
										"406 NOT_ACCEPTABLE",
										"407 PROXY_AUTHENTICATION_REQUIRED",
										"408 REQUEST_TIMEOUT",
										"409 CONFLICT",
										"410 GONE",
										"411 LENGTH_REQUIRED",
										"412 PRECONDITION_FAILED",
										"413 PAYLOAD_TOO_LARGE",
										"413 REQUEST_ENTITY_TOO_LARGE",
										"414 URI_TOO_LONG",
										"414 REQUEST_URI_TOO_LONG",
										"415 UNSUPPORTED_MEDIA_TYPE",
										"416 REQUESTED_RANGE_NOT_SATISFIABLE",
										"417 EXPECTATION_FAILED",
										"418 I_AM_A_TEAPOT",
										"419 INSUFFICIENT_SPACE_ON_RESOURCE",
										"420 METHOD_FAILURE",
										"421 DESTINATION_LOCKED",
										"422 UNPROCESSABLE_ENTITY",
										"423 LOCKED",
										"424 FAILED_DEPENDENCY",
										"425 TOO_EARLY",
										"426 UPGRADE_REQUIRED",
										"428 PRECONDITION_REQUIRED",
										"429 TOO_MANY_REQUESTS",
										"431 REQUEST_HEADER_FIELDS_TOO_LARGE",
										"451 UNAVAILABLE_FOR_LEGAL_REASONS",
										"500 INTERNAL_SERVER_ERROR",
										"501 NOT_IMPLEMENTED",
										"502 BAD_GATEWAY",
										"503 SERVICE_UNAVAILABLE",
										"504 GATEWAY_TIMEOUT",
										"505 HTTP_VERSION_NOT_SUPPORTED",
										"506 VARIANT_ALSO_NEGOTIATES",
										"507 INSUFFICIENT_STORAGE",
										"508 LOOP_DETECTED",
										"509 BANDWIDTH_LIMIT_EXCEEDED",
										"510 NOT_EXTENDED",
										"511 NETWORK_AUTHENTICATION_REQUIRED"
									]
								}
							}
						}
					}
				}
			}
		},
		"/api/tutorials": {
			"get": {
				"tags": [
					"filter",
					"Tutorial",
					"tutorials",
					"get"
				],
				"summary": "View Tutorials",
				"description": "Filters Tutorials by title or description. If no title or description defined, returns all tutorials.",
				"operationId": "getTutorials",
				"parameters": [
					{
						"name": "title",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "description",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TutorialDto"
								}
							}
						}
					},
					"204": {
						"description": "No Tutorials found",
						"content": {
							"*/*": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/TutorialDto"
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"*/*": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/TutorialDto"
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": [
					"Tutorial",
					"post",
					"tutorials"
				],
				"summary": "Create a new Tutorial",
				"description": "Creates a new tutorial from provided data or an empty tutorial if no data is provided.",
				"operationId": "createTutorial",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TutorialDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Tutorial successfully created",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/TutorialDto"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/TutorialDto"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": [
					"Tutorial",
					"tutorials",
					"delete"
				],
				"summary": "Delete all Tutorials",
				"description": "Removes all tutorials from the database.",
				"operationId": "deleteAllTutorials",
				"responses": {
					"204": {
						"description": "All tutorials deleted",
						"content": {
							"*/*": {
								"schema": {
									"type": "string",
									"enum": [
										"100 CONTINUE",
										"101 SWITCHING_PROTOCOLS",
										"102 PROCESSING",
										"103 EARLY_HINTS",
										"103 CHECKPOINT",
										"200 OK",
										"201 CREATED",
										"202 ACCEPTED",
										"203 NON_AUTHORITATIVE_INFORMATION",
										"204 NO_CONTENT",
										"205 RESET_CONTENT",
										"206 PARTIAL_CONTENT",
										"207 MULTI_STATUS",
										"208 ALREADY_REPORTED",
										"226 IM_USED",
										"300 MULTIPLE_CHOICES",
										"301 MOVED_PERMANENTLY",
										"302 FOUND",
										"302 MOVED_TEMPORARILY",
										"303 SEE_OTHER",
										"304 NOT_MODIFIED",
										"305 USE_PROXY",
										"307 TEMPORARY_REDIRECT",
										"308 PERMANENT_REDIRECT",
										"400 BAD_REQUEST",
										"401 UNAUTHORIZED",
										"402 PAYMENT_REQUIRED",
										"403 FORBIDDEN",
										"404 NOT_FOUND",
										"405 METHOD_NOT_ALLOWED",
										"406 NOT_ACCEPTABLE",
										"407 PROXY_AUTHENTICATION_REQUIRED",
										"408 REQUEST_TIMEOUT",
										"409 CONFLICT",
										"410 GONE",
										"411 LENGTH_REQUIRED",
										"412 PRECONDITION_FAILED",
										"413 PAYLOAD_TOO_LARGE",
										"413 REQUEST_ENTITY_TOO_LARGE",
										"414 URI_TOO_LONG",
										"414 REQUEST_URI_TOO_LONG",
										"415 UNSUPPORTED_MEDIA_TYPE",
										"416 REQUESTED_RANGE_NOT_SATISFIABLE",
										"417 EXPECTATION_FAILED",
										"418 I_AM_A_TEAPOT",
										"419 INSUFFICIENT_SPACE_ON_RESOURCE",
										"420 METHOD_FAILURE",
										"421 DESTINATION_LOCKED",
										"422 UNPROCESSABLE_ENTITY",
										"423 LOCKED",
										"424 FAILED_DEPENDENCY",
										"425 TOO_EARLY",
										"426 UPGRADE_REQUIRED",
										"428 PRECONDITION_REQUIRED",
										"429 TOO_MANY_REQUESTS",
										"431 REQUEST_HEADER_FIELDS_TOO_LARGE",
										"451 UNAVAILABLE_FOR_LEGAL_REASONS",
										"500 INTERNAL_SERVER_ERROR",
										"501 NOT_IMPLEMENTED",
										"502 BAD_GATEWAY",
										"503 SERVICE_UNAVAILABLE",
										"504 GATEWAY_TIMEOUT",
										"505 HTTP_VERSION_NOT_SUPPORTED",
										"506 VARIANT_ALSO_NEGOTIATES",
										"507 INSUFFICIENT_STORAGE",
										"508 LOOP_DETECTED",
										"509 BANDWIDTH_LIMIT_EXCEEDED",
										"510 NOT_EXTENDED",
										"511 NETWORK_AUTHENTICATION_REQUIRED"
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"*/*": {
								"schema": {
									"type": "string",
									"enum": [
										"100 CONTINUE",
										"101 SWITCHING_PROTOCOLS",
										"102 PROCESSING",
										"103 EARLY_HINTS",
										"103 CHECKPOINT",
										"200 OK",
										"201 CREATED",
										"202 ACCEPTED",
										"203 NON_AUTHORITATIVE_INFORMATION",
										"204 NO_CONTENT",
										"205 RESET_CONTENT",
										"206 PARTIAL_CONTENT",
										"207 MULTI_STATUS",
										"208 ALREADY_REPORTED",
										"226 IM_USED",
										"300 MULTIPLE_CHOICES",
										"301 MOVED_PERMANENTLY",
										"302 FOUND",
										"302 MOVED_TEMPORARILY",
										"303 SEE_OTHER",
										"304 NOT_MODIFIED",
										"305 USE_PROXY",
										"307 TEMPORARY_REDIRECT",
										"308 PERMANENT_REDIRECT",
										"400 BAD_REQUEST",
										"401 UNAUTHORIZED",
										"402 PAYMENT_REQUIRED",
										"403 FORBIDDEN",
										"404 NOT_FOUND",
										"405 METHOD_NOT_ALLOWED",
										"406 NOT_ACCEPTABLE",
										"407 PROXY_AUTHENTICATION_REQUIRED",
										"408 REQUEST_TIMEOUT",
										"409 CONFLICT",
										"410 GONE",
										"411 LENGTH_REQUIRED",
										"412 PRECONDITION_FAILED",
										"413 PAYLOAD_TOO_LARGE",
										"413 REQUEST_ENTITY_TOO_LARGE",
										"414 URI_TOO_LONG",
										"414 REQUEST_URI_TOO_LONG",
										"415 UNSUPPORTED_MEDIA_TYPE",
										"416 REQUESTED_RANGE_NOT_SATISFIABLE",
										"417 EXPECTATION_FAILED",
										"418 I_AM_A_TEAPOT",
										"419 INSUFFICIENT_SPACE_ON_RESOURCE",
										"420 METHOD_FAILURE",
										"421 DESTINATION_LOCKED",
										"422 UNPROCESSABLE_ENTITY",
										"423 LOCKED",
										"424 FAILED_DEPENDENCY",
										"425 TOO_EARLY",
										"426 UPGRADE_REQUIRED",
										"428 PRECONDITION_REQUIRED",
										"429 TOO_MANY_REQUESTS",
										"431 REQUEST_HEADER_FIELDS_TOO_LARGE",
										"451 UNAVAILABLE_FOR_LEGAL_REASONS",
										"500 INTERNAL_SERVER_ERROR",
										"501 NOT_IMPLEMENTED",
										"502 BAD_GATEWAY",
										"503 SERVICE_UNAVAILABLE",
										"504 GATEWAY_TIMEOUT",
										"505 HTTP_VERSION_NOT_SUPPORTED",
										"506 VARIANT_ALSO_NEGOTIATES",
										"507 INSUFFICIENT_STORAGE",
										"508 LOOP_DETECTED",
										"509 BANDWIDTH_LIMIT_EXCEEDED",
										"510 NOT_EXTENDED",
										"511 NETWORK_AUTHENTICATION_REQUIRED"
									]
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"TutorialDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"format": "int64"
					},
					"title": {
						"type": "string",
						"minLength": 1
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"title"
				]
			}
		}
	}
}